requirements.txt
IPy
paramiko
PyYAML
PyQt5